Fix the integration test.
This fixes timers to actually be disabled when passing a delay of 0.

It also change NetworkManager::receive to return false if the packet has
recently been seen (since it will already have been processed). Without
this, a node will receive the rebroadcasted version of its own packets
and get confused.

Change-Id: I9eaf40f59047592fa129ede896767b8dd48b1217