WIP enable the watchdog timer.
This enables a timer which will reset the board if it has not 'fed' the
watchdog in ~100ms. The main loop feeds the watchdog, so the board
should reset if it hangs.

This also updates the bootloader to disable the watchdog. This is very
important - without this update, a board with a bad program may be
unprogrammable over USB until an updated bootloader is flashed via
openocd.

Change-Id: Ia8e93f49543822d839f89a6a51f101713bd44310